+++ bin/store/src/models/account_phone_number_model.rs
@@ -8,11 +8,6 @@ use serde::{Deserialize, Serialize};
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[serde(default)]
 pub struct AccountPhoneNumberModel {
-    pub account_profile_id: Option<String>,
-    pub raw_phone_number: Option<String>,
-    pub is_primary: Option<bool>,
-    pub iso_code: Option<String>,
-    pub country_code: Option<String>,
     pub tombstone: Option<i32>,
     pub status: Option<String>,
     pub previous_status: Option<String>,
@@ -34,4 +29,9 @@ pub struct AccountPhoneNumberModel {
     pub sensitivity_level: Option<i32>,
     pub sync_status: Option<String>,
     pub is_batch: Option<bool>,
+    pub account_profile_id: Option<String>,
+    pub raw_phone_number: Option<String>,
+    pub is_primary: Option<bool>,
+    pub iso_code: Option<String>,
+    pub country_code: Option<String>,
 }